EdDSA: add optional signature verification to prevent outputting faulty signatures
EdDSA is known to be vulnerable to fault attacks which can lead to secret key extraction if
two signatures over the same message can be collected.
Randomly occurring bitflips in specific parts of the computation might in principle result
in vulnerable faulty signatures being generated, hence we add the option to validate the signatures
before outputting them.